octal number conversions

+++ include/LinkedList.h
@@ -2,11 +2,11 @@
 
 class Node {
 public:
-  int info;
+  unsigned long long info;
   Node *next;
 
   // Initializes a node
-  Node(int info, Node *next) : info(info), next(next) {}
+  Node(unsigned long long info, Node *next) : info(info), next(next) {}
   Node() : next(nullptr) {}
 };
 
@@ -25,38 +25,38 @@ public:
   ~LinkedList(){};
 
   // Returns the value stored at head
-  int getHeadInfo();
+  unsigned long long getHeadInfo();
 
   // Returns the value stored at tail
-  int getTailInfo();
+  unsigned long long getTailInfo();
 
   // Checks if the list is empty or not
   bool isEmpty();
 
   // Adds data at the top(head) of list
-  void addToHead(int data);
+  void addToHead(unsigned long long data);
 
   // Adds data at the bottom(tail) of list
-  void addToTail(int data);
+  void addToTail(unsigned long long data);
 
   // Goes through every single elements(if present) of the list
   void traverse(char seperator = ' ');
 
   // Adds data next to the given predecessor
-  void add(int data, Node *predecessor);
+  void add(unsigned long long data, Node *predecessor);
 
   // Removes and returns data from the top(head)
-  int removeFromHead();
+  unsigned long long removeFromHead();
 
   // Removes and returns data from the bottom(tail)
-  int removeFromTail();
+  unsigned long long removeFromTail();
 
   // Removes the given data from the list
-  void remove(int data);
+  void remove(unsigned long long data);
 
   // Searches if the given data exists in the list
-  bool search(int data);
+  bool search(unsigned long long data);
 
   // Searches and returns the node which contains the given data
-  Node *retrive(int data, Node *dataOutPointer);
+  Node *retrive(unsigned long long data, Node *dataOutPointer);
 };
